Stage32: normalize integral coset class keys

+++ stages/stage32/residual-32-01-production/direct_picard_integral_coset_bound.py
@@ -130,7 +130,7 @@ class DirectPicardIntegralCosetLowerBound:
 
         def class_key(residue: tuple[int, int, int]) -> tuple[int, ...]:
             return tuple(
-                sum(R[i, j] * residue[j] for j in range(3)) % common_den
+                int(sum(R[i, j] * residue[j] for j in range(3)) % common_den)
                 for i in range(61)
             )
